Added KML file support
Use GPSBabel to load KML files in GpxFile objects. Other file types
recognised by GPSBabel can be added easily.

Note that currently, there is no check whether GPSBabel is really
present, so the gpsbabel binary must be present in the executable
search path. Also, the file dialog does is not aware of KML files. This
is scheduled for the future.

Signed-off-by: Jens Lang <jenslang@users.sf.net>